tested seed file and updated scripts

+++ server/db/seed.ts
@@ -1,24 +1,32 @@
 import db from "./db";
 import User from "./User";
 
+import dotenv from 'dotenv'
+dotenv.config()
+
 const seed = async () => {
-  await db.sync({ force: true });
+    console.log('STARTING SEED')
+    console.log("SEED VALUE", process.env.SEED)
+    if (process.env.SEED === 'true') {
+      await db.sync({ force: true });
+    
+      const [olivia, lucy, joe, bobby] = await Promise.all([
+        User.create({ username: "olivia", password: "123", email: 'ocjarman@gmail.com' }),
+        User.create({ username: "lucy", password: "123" }),
+        User.create({ username: "joe", password: "123" }),
+        User.create({ username: "liv", password: "123" }),
+      ]);
 
-  const [moe, lucy, larry, ethyl] = await Promise.all([
-    User.create({ username: "olivia", password: "123" }),
-    User.create({ username: "lucy", password: "123" }),
-    User.create({ username: "larry", password: "123" }),
-    User.create({ username: "ethyl", password: "123" }),
-  ]);
-
-  return {
-    users: {
-      moe,
-      lucy,
-      larry,
-      ethyl,
-    },
-  };
+      console.log('ENDING SEED')
+      return {
+        users: {
+          olivia,
+          lucy,
+          joe,
+          bobby,
+        },
+      };
+    }
 };
 
-export default seed();
+seed()
